+++ CHANGELOG.md
@@ -41,6 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improvements
 
+- [#702](https://github.com/alleslabs/celatone-frontend/pull/702) api v1 - contract's query msgs
 - [#696](https://github.com/alleslabs/celatone-frontend/pull/696) api v1 - block details
 - [#695](https://github.com/alleslabs/celatone-frontend/pull/695) api v1 - contract states
 - [#678](https://github.com/alleslabs/celatone-frontend/pull/678) api v1 - contract table counts